Refactor methods into individual files
